refactor(backend): split catalog install boundaries
