+++ apps/backend/src/catalog/distribution/install/install.ts
@@ -1,1374 +1,9 @@
-import { randomUUID } from "node:crypto";
-import { z } from "zod";
-import {
-  transactionWithWorkspaceScope,
-  transactionWithWorkspaceScopeReadOnly,
-  type DatabaseExecutor,
-} from "../../../database";
-import type { CardMetadata, CardSourceMetadata } from "../../../cards/types";
-import { normalizeCardMetadata } from "../../../cards/shared";
-import {
-  isValidMediaAssetLastOperationId,
-  isValidMediaAssetLastOperationIdPrefix,
-  maximumMediaAssetLastOperationIdLength,
-} from "../../../mediaAssets/lastOperationId";
-import { assertReplicaBelongsToWorkspaceInExecutor } from "../../../mediaAssets/workspaceReplicas";
-import { HttpError } from "../../../shared/errors";
-import {
-  buildSuggestedCardImportTag,
-  normalizeCardImportTagOptions,
-  planCardImportTags,
-  type CardImportTagPlan,
-} from "../../../shared/cardImportTags";
-import { normalizeIsoTimestamp } from "../../../sync/conflicts/lww";
-import {
-  insertSyncChange,
-  lockWorkspaceSyncMetadataForHotChangesInExecutor,
-  type HotChangeWriteLock,
-} from "../../../sync/replication/changes";
-import { rewriteMarkdownFcAssetUrlsToFcAssets } from "../../../workspacePackages";
-import {
-  isLowercaseWorkspaceId,
-  normalizeWorkspaceId,
-} from "../../../workspaces/identity";
-import {
-  normalizeNonEmptyString,
-  toIsoString,
-  toOptionalIsoString,
-  toSafeNumber,
-} from "../../common";
-import type {
-  CatalogPackageStatus,
-  TimestampValue,
-  CatalogPackageInstallConfirmInput,
-  CatalogPackageInstallPreview,
-  CatalogPackageInstallPreviewInput,
-  CatalogPackageInstallResult,
-  CatalogPackageInstallPackageVersion,
-  CatalogInstalledCard,
-  CatalogInstalledMediaAsset,
-} from "../../types";
+export {
+  catalogPackageInstallOperationIdPrefixMaximumLength,
+  installCatalogPackageVersion,
+  installCatalogPackageVersionInExecutor,
+  isValidCatalogPackageInstallOperationIdPrefix,
+  previewCatalogPackageInstall,
+  previewCatalogPackageInstallInExecutor,
+} from "./index";
 
-const uuidPattern = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/iu;
-const maximumJavaScriptArrayIndex = 4_294_967_294;
-const catalogPackageInstallMediaLastOperationIdMaximumSuffix =
-  `:media:${maximumJavaScriptArrayIndex}`;
-const catalogPackageInstallCardLastOperationIdMaximumSuffix =
-  `:card:${maximumJavaScriptArrayIndex}`;
-const catalogPackageInstallLastOperationIdMaximumSuffixLength = Math.max(
-  catalogPackageInstallMediaLastOperationIdMaximumSuffix.length,
-  catalogPackageInstallCardLastOperationIdMaximumSuffix.length,
-);
-
-export const catalogPackageInstallOperationIdPrefixMaximumLength =
-  maximumMediaAssetLastOperationIdLength
-  - catalogPackageInstallLastOperationIdMaximumSuffixLength;
-
-type CatalogPackageInstallVersionRow = Readonly<{
-  package_version_id: string;
-  package_id: string;
-  version_number: string | number;
-  status: CatalogPackageStatus;
-  slug: string;
-  title: string;
-  summary: string;
-  description: string;
-  language_tags: ReadonlyArray<string>;
-  topic_tags: ReadonlyArray<string>;
-  license: string;
-  content_warning: string | null;
-  cover_package_media_key: string | null;
-  card_count: string | number;
-  created_at: TimestampValue;
-  published_at: TimestampValue | null;
-  author_id: string;
-  author_slug: string;
-  author_display_name: string;
-}>;
-
-type CatalogPackageInstallMediaAssetRow = Readonly<{
-  package_media_asset_id: string;
-  package_media_key: string;
-  media_blob_id: string;
-}>;
-
-type CatalogPackageInstallCardRow = Readonly<{
-  package_card_id: string;
-  stable_card_key: string;
-  ordinal: string | number;
-  front_text: string;
-  back_text: string;
-  card_type: string;
-  metadata: CardMetadata;
-  tags: ReadonlyArray<string>;
-  media_asset_keys: ReadonlyArray<string>;
-}>;
-
-type CatalogPackageInstallTagRow = Readonly<{
-  tags: ReadonlyArray<string>;
-}>;
-
-type CatalogPackageInstallOperationConflictRow = Readonly<{
-  entity_type: string;
-  entity_id: string;
-  last_operation_id: string;
-}>;
-
-type CatalogPackageInstallIdempotencyRow = Readonly<{
-  package_version_id: string;
-  installed_at: TimestampValue;
-  client_updated_at: TimestampValue;
-  last_modified_by_replica_id: string;
-  operation_id_prefix: string;
-  add_import_tag: boolean;
-  import_tag: string | null;
-  remove_tags: ReadonlyArray<string>;
-  install_result: unknown;
-}>;
-
-type NormalizedCatalogPackageInstallConfirmInput = Readonly<{
-  installId: string;
-  installedAt: string;
-  clientUpdatedAt: string;
-  lastModifiedByReplicaId: string;
-  operationIdPrefix: string;
-  addImportTag: boolean;
-  importTag: string;
-  removeTags: ReadonlyArray<string>;
-}>;
-
-type CatalogPackageInstallRequestIdentity = Readonly<{
-  packageVersionId: string;
-  installedAt: string;
-  clientUpdatedAt: string;
-  lastModifiedByReplicaId: string;
-  operationIdPrefix: string;
-  addImportTag: boolean;
-  importTag: string | null;
-  removeTags: ReadonlyArray<string>;
-}>;
-
-const catalogPackageInstallUuidSchema = z.string().uuid();
-const catalogPackageInstallDateTimeSchema = z.string().datetime({ offset: true });
-
-const catalogPackageInstallAuthorSchema = z.object({
-  authorId: catalogPackageInstallUuidSchema,
-  slug: z.string(),
-  displayName: z.string(),
-}).strict();
-
-const catalogPackageInstallPackageVersionSchema = z.object({
-  packageVersionId: catalogPackageInstallUuidSchema,
-  packageId: catalogPackageInstallUuidSchema,
-  versionNumber: z.number().int().positive(),
-  slug: z.string(),
-  title: z.string(),
-  summary: z.string(),
-  description: z.string(),
-  languageTags: z.array(z.string()),
-  topicTags: z.array(z.string()),
-  license: z.string(),
-  contentWarning: z.string().nullable(),
-  coverPackageMediaKey: z.string().nullable(),
-  cardCount: z.number().int().nonnegative(),
-  createdAt: catalogPackageInstallDateTimeSchema,
-  publishedAt: catalogPackageInstallDateTimeSchema.nullable(),
-  author: catalogPackageInstallAuthorSchema,
-}).strict();
-
-const catalogPackageInstallResultSchema = z.object({
-  packageVersion: catalogPackageInstallPackageVersionSchema,
-  installedCards: z.array(z.object({
-    packageCardId: catalogPackageInstallUuidSchema,
-    stableCardKey: z.string(),
-    ordinal: z.number().int().positive(),
-    cardId: catalogPackageInstallUuidSchema,
-  }).strict()),
-  installedMediaAssets: z.array(z.object({
-    packageMediaAssetId: catalogPackageInstallUuidSchema,
-    packageMediaKey: z.string(),
-    mediaAssetId: catalogPackageInstallUuidSchema,
-  }).strict()),
-  summary: z.object({
-    cardCount: z.number().int().nonnegative(),
-    mediaAssetCount: z.number().int().nonnegative(),
-    installId: z.string(),
-    installedAt: catalogPackageInstallDateTimeSchema,
-    keptTagCount: z.number().int().nonnegative(),
-    removedTagCount: z.number().int().nonnegative(),
-    importTag: z.string().nullable(),
-  }).strict(),
-}).strict();
-
-const catalogPackageInstallVersionColumns = [
-  "package_versions.package_version_id AS package_version_id",
-  "package_versions.package_id AS package_id",
-  "package_versions.version_number AS version_number",
-  "package_versions.status AS status",
-  "package_versions.slug AS slug",
-  "package_versions.title AS title",
-  "package_versions.summary AS summary",
-  "package_versions.description AS description",
-  "package_versions.language_tags AS language_tags",
-  "package_versions.topic_tags AS topic_tags",
-  "package_versions.license AS license",
-  "package_versions.content_warning AS content_warning",
-  "package_versions.cover_package_media_key AS cover_package_media_key",
-  "package_versions.card_count AS card_count",
-  "package_versions.created_at AS created_at",
-  "package_versions.published_at AS published_at",
-  "authors.author_id AS author_id",
-  "authors.slug AS author_slug",
-  "authors.display_name AS author_display_name",
-].join(", ");
-
-function normalizeUuidString(value: string, fieldName: string): string {
-  const normalizedValue = normalizeNonEmptyString(value, fieldName).toLowerCase();
-  if (uuidPattern.test(normalizedValue) === false) {
-    throw new HttpError(400, `${fieldName} must be a UUID`, "CATALOG_PACKAGE_INSTALL_INVALID_INPUT");
-  }
-
-  return normalizedValue;
-}
-
-function normalizeCatalogWorkspaceId(value: string): string {
-  const normalizedValue = normalizeWorkspaceId(
-    normalizeNonEmptyString(value, "workspaceId"),
-  ).toLowerCase();
-  if (isLowercaseWorkspaceId(normalizedValue) === false) {
-    throw new HttpError(
-      400,
-      "workspaceId must be a UUID",
-      "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-    );
-  }
-
-  return normalizedValue;
-}
-
-function normalizeBoundedNonEmptyString(value: string, fieldName: string, maximumLength: number): string {
-  const normalizedValue = normalizeNonEmptyString(value, fieldName);
-  if (normalizedValue.length > maximumLength) {
-    throw new HttpError(
-      400,
-      `${fieldName} must contain at most ${maximumLength} characters`,
-      "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-    );
-  }
-
-  return normalizedValue;
-}
-
-export function isValidCatalogPackageInstallOperationIdPrefix(
-  value: string,
-): boolean {
-  return isValidMediaAssetLastOperationIdPrefix(
-    value,
-    catalogPackageInstallOperationIdPrefixMaximumLength,
-  );
-}
-
-function normalizeCatalogPackageInstallOperationIdPrefix(value: string): string {
-  if (isValidCatalogPackageInstallOperationIdPrefix(value)) {
-    return value;
-  }
-
-  throw new HttpError(
-    400,
-    [
-      "operationIdPrefix must be",
-      `1 to ${catalogPackageInstallOperationIdPrefixMaximumLength}`,
-      "printable ASCII characters without leading or trailing spaces.",
-    ].join(" "),
-    "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-  );
-}
-
-function normalizeCatalogInstallIsoTimestamp(value: string, fieldName: string): string {
-  try {
-    return normalizeIsoTimestamp(value, fieldName);
-  } catch {
-    throw new HttpError(
-      400,
-      `${fieldName} must be a valid ISO timestamp`,
-      "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-    );
-  }
-}
-
-function normalizeCatalogPackageInstallConfirmInput(
-  input: CatalogPackageInstallConfirmInput,
-): NormalizedCatalogPackageInstallConfirmInput {
-  let normalizedTagOptions: ReturnType<typeof normalizeCardImportTagOptions>;
-  try {
-    normalizedTagOptions = normalizeCardImportTagOptions({
-      addImportTag: input.addImportTag === undefined ? false : input.addImportTag,
-      importTag: input.importTag === undefined ? "" : input.importTag,
-      removeTags: input.removeTags === undefined ? [] : input.removeTags,
-    });
-  } catch (error) {
-    throw new HttpError(
-      400,
-      `Catalog package install tag options are invalid. reason=${error instanceof Error ? error.message : String(error)}`,
-      "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-    );
-  }
-
-  return {
-    installId: normalizeBoundedNonEmptyString(input.installId, "installId", 128),
-    installedAt: normalizeCatalogInstallIsoTimestamp(input.installedAt, "installedAt"),
-    clientUpdatedAt: normalizeCatalogInstallIsoTimestamp(input.clientUpdatedAt, "clientUpdatedAt"),
-    lastModifiedByReplicaId: normalizeUuidString(input.lastModifiedByReplicaId, "lastModifiedByReplicaId"),
-    operationIdPrefix: normalizeCatalogPackageInstallOperationIdPrefix(input.operationIdPrefix),
-    addImportTag: normalizedTagOptions.addImportTag,
-    importTag: normalizedTagOptions.importTag,
-    removeTags: normalizedTagOptions.removeTags,
-  };
-}
-
-function createCatalogPackageInstallRequestIdentity(
-  packageVersionId: string,
-  input: NormalizedCatalogPackageInstallConfirmInput,
-): CatalogPackageInstallRequestIdentity {
-  return {
-    packageVersionId,
-    installedAt: input.installedAt,
-    clientUpdatedAt: input.clientUpdatedAt,
-    lastModifiedByReplicaId: input.lastModifiedByReplicaId,
-    operationIdPrefix: input.operationIdPrefix,
-    addImportTag: input.addImportTag,
-    importTag: input.addImportTag ? input.importTag : null,
-    removeTags: [...input.removeTags].sort(),
-  };
-}
-
-function parseCatalogPackageInstallStoredResult(
-  value: unknown,
-  workspaceId: string,
-  installId: string,
-): CatalogPackageInstallResult {
-  const parsedResult = catalogPackageInstallResultSchema.safeParse(value);
-  if (parsedResult.success === false) {
-    throw new HttpError(
-      500,
-      [
-        "Stored catalog package install result cannot be replayed safely.",
-        `workspaceId=${workspaceId}`,
-        `installId=${installId}`,
-        `reason=${parsedResult.error.message}`,
-        "Repair the catalog install idempotency record before retrying.",
-      ].join(" "),
-      "CATALOG_PACKAGE_INSTALL_STORED_RESULT_INVALID",
-    );
-  }
-
-  return parsedResult.data;
-}
-
-function catalogPackageInstallStoredResultMismatchFields(
-  row: CatalogPackageInstallIdempotencyRow,
-  identity: CatalogPackageInstallRequestIdentity,
-  result: CatalogPackageInstallResult,
-  installId: string,
-): ReadonlyArray<string> {
-  const mismatchFields: Array<string> = [];
-  if (result.summary.installId !== installId) mismatchFields.push("summary.installId");
-  if (
-    result.packageVersion.packageVersionId !== row.package_version_id
-    || result.packageVersion.packageVersionId !== identity.packageVersionId
-  ) {
-    mismatchFields.push("packageVersion.packageVersionId");
-  }
-  if (
-    result.summary.installedAt !== toIsoString(row.installed_at)
-    || result.summary.installedAt !== identity.installedAt
-  ) {
-    mismatchFields.push("summary.installedAt");
-  }
-  if (result.summary.importTag !== row.import_tag || result.summary.importTag !== identity.importTag) {
-    mismatchFields.push("summary.importTag");
-  }
-  if (result.summary.cardCount !== result.installedCards.length) mismatchFields.push("summary.cardCount");
-  if (result.packageVersion.cardCount !== result.installedCards.length) {
-    mismatchFields.push("packageVersion.cardCount");
-  }
-  if (result.summary.mediaAssetCount !== result.installedMediaAssets.length) {
-    mismatchFields.push("summary.mediaAssetCount");
-  }
-  if (result.summary.removedTagCount !== row.remove_tags.length) {
-    mismatchFields.push("summary.removedTagCount");
-  }
-  return mismatchFields;
-}
-
-function assertCatalogPackageInstallStoredResultMatchesIdentity(
-  row: CatalogPackageInstallIdempotencyRow,
-  identity: CatalogPackageInstallRequestIdentity,
-  result: CatalogPackageInstallResult,
-  workspaceId: string,
-  installId: string,
-): void {
-  const mismatchFields = catalogPackageInstallStoredResultMismatchFields(
-    row,
-    identity,
-    result,
-    installId,
-  );
-  if (mismatchFields.length === 0) {
-    return;
-  }
-
-  throw new HttpError(
-    500,
-    [
-      "Stored catalog package install result does not match its durable request identity.",
-      `workspaceId=${workspaceId}`,
-      `installId=${installId}`,
-      `mismatchedFields=${mismatchFields.join(",")}`,
-      "Repair the catalog install idempotency record before retrying.",
-    ].join(" "),
-    "CATALOG_PACKAGE_INSTALL_STORED_RESULT_INVALID",
-  );
-}
-
-function catalogPackageInstallRequestIdentityMismatchFields(
-  row: CatalogPackageInstallIdempotencyRow,
-  identity: CatalogPackageInstallRequestIdentity,
-): ReadonlyArray<string> {
-  const mismatchFields: Array<string> = [];
-  if (row.package_version_id !== identity.packageVersionId) mismatchFields.push("packageVersionId");
-  if (toIsoString(row.installed_at) !== identity.installedAt) mismatchFields.push("installedAt");
-  if (toIsoString(row.client_updated_at) !== identity.clientUpdatedAt) mismatchFields.push("clientUpdatedAt");
-  if (row.last_modified_by_replica_id !== identity.lastModifiedByReplicaId) {
-    mismatchFields.push("lastModifiedByReplicaId");
-  }
-  if (row.operation_id_prefix !== identity.operationIdPrefix) mismatchFields.push("operationIdPrefix");
-  if (row.add_import_tag !== identity.addImportTag) mismatchFields.push("addImportTag");
-  if (row.import_tag !== identity.importTag) mismatchFields.push("importTag");
-  if (JSON.stringify(row.remove_tags) !== JSON.stringify(identity.removeTags)) mismatchFields.push("removeTags");
-  return mismatchFields;
-}
-
-async function loadCatalogPackageInstallReplayInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  installId: string,
-  identity: CatalogPackageInstallRequestIdentity,
-): Promise<CatalogPackageInstallResult | null> {
-  const result = await executor.query<CatalogPackageInstallIdempotencyRow>(
-    [
-      "SELECT package_version_id::text, installed_at, client_updated_at,",
-      "last_modified_by_replica_id::text, operation_id_prefix, add_import_tag,",
-      "import_tag, remove_tags, install_result",
-      "FROM sync.catalog_package_install_idempotency",
-      "WHERE workspace_id = $1 AND install_id = $2",
-    ].join(" "),
-    [workspaceId, installId],
-  );
-  const row = result.rows[0];
-  if (row === undefined) {
-    return null;
-  }
-
-  const mismatchFields = catalogPackageInstallRequestIdentityMismatchFields(row, identity);
-  if (mismatchFields.length !== 0) {
-    throw new HttpError(
-      409,
-      [
-        "Catalog package install idempotency key was already used with a different normalized request.",
-        `workspaceId=${workspaceId}`,
-        `installId=${installId}`,
-        `mismatchedFields=${mismatchFields.join(",")}`,
-        "Use a new installId for an explicit repeat import.",
-      ].join(" "),
-      "CATALOG_PACKAGE_INSTALL_IDEMPOTENCY_CONFLICT",
-    );
-  }
-
-  const storedResult = parseCatalogPackageInstallStoredResult(row.install_result, workspaceId, installId);
-  assertCatalogPackageInstallStoredResultMatchesIdentity(
-    row,
-    identity,
-    storedResult,
-    workspaceId,
-    installId,
-  );
-
-  return storedResult;
-}
-
-async function insertCatalogPackageInstallIdempotencyResultInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  installId: string,
-  identity: CatalogPackageInstallRequestIdentity,
-  result: CatalogPackageInstallResult,
-): Promise<void> {
-  await executor.query(
-    [
-      "INSERT INTO sync.catalog_package_install_idempotency",
-      "(workspace_id, install_id, package_version_id, installed_at, client_updated_at,",
-      "last_modified_by_replica_id, operation_id_prefix, add_import_tag, import_tag, remove_tags, install_result)",
-      "VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11::jsonb)",
-    ].join(" "),
-    [
-      workspaceId,
-      installId,
-      identity.packageVersionId,
-      identity.installedAt,
-      identity.clientUpdatedAt,
-      identity.lastModifiedByReplicaId,
-      identity.operationIdPrefix,
-      identity.addImportTag,
-      identity.importTag,
-      identity.removeTags,
-      JSON.stringify(result),
-    ],
-  );
-}
-
-function assertCatalogPackageVersionIsPublished(row: CatalogPackageInstallVersionRow): void {
-  if (row.status === "published") {
-    return;
-  }
-
-  throw new HttpError(
-    409,
-    `Catalog package version must be published before installation. packageVersionId=${row.package_version_id} status=${row.status}`,
-    "CATALOG_PACKAGE_VERSION_NOT_PUBLISHED",
-  );
-}
-
-function mapCatalogPackageInstallPackageVersion(
-  row: CatalogPackageInstallVersionRow,
-): CatalogPackageInstallPackageVersion {
-  return {
-    packageVersionId: row.package_version_id,
-    packageId: row.package_id,
-    versionNumber: toSafeNumber(row.version_number, "version_number"),
-    slug: row.slug,
-    title: row.title,
-    summary: row.summary,
-    description: row.description,
-    languageTags: [...row.language_tags],
-    topicTags: [...row.topic_tags],
-    license: row.license,
-    contentWarning: row.content_warning,
-    coverPackageMediaKey: row.cover_package_media_key,
-    cardCount: toSafeNumber(row.card_count, "card_count"),
-    createdAt: toIsoString(row.created_at),
-    publishedAt: toOptionalIsoString(row.published_at),
-    author: {
-      authorId: row.author_id,
-      slug: row.author_slug,
-      displayName: row.author_display_name,
-    },
-  };
-}
-
-async function loadCatalogPackageInstallVersionInExecutor(
-  executor: DatabaseExecutor,
-  packageVersionId: string,
-): Promise<CatalogPackageInstallVersionRow> {
-  const result = await executor.query<CatalogPackageInstallVersionRow>(
-    [
-      "SELECT",
-      catalogPackageInstallVersionColumns,
-      "FROM catalog.package_versions AS package_versions",
-      "INNER JOIN catalog.packages AS packages",
-      "ON packages.package_id = package_versions.package_id",
-      "INNER JOIN catalog.authors AS authors",
-      "ON authors.author_id = packages.author_id",
-      "WHERE package_versions.package_version_id = $1",
-    ].join(" "),
-    [packageVersionId],
-  );
-  const row = result.rows[0];
-  if (row === undefined) {
-    throw new HttpError(
-      404,
-      `Catalog package version not found. packageVersionId=${packageVersionId}`,
-      "CATALOG_PACKAGE_VERSION_NOT_FOUND",
-    );
-  }
-
-  assertCatalogPackageVersionIsPublished(row);
-  return row;
-}
-
-async function loadCatalogPackageInstallVersionForInstallInExecutor(
-  executor: DatabaseExecutor,
-  packageVersionId: string,
-): Promise<CatalogPackageInstallVersionRow> {
-  const result = await executor.query<CatalogPackageInstallVersionRow>(
-    [
-      "SELECT",
-      catalogPackageInstallVersionColumns,
-      "FROM catalog.package_versions AS package_versions",
-      "INNER JOIN catalog.packages AS packages",
-      "ON packages.package_id = package_versions.package_id",
-      "INNER JOIN catalog.authors AS authors",
-      "ON authors.author_id = packages.author_id",
-      "WHERE package_versions.package_version_id = $1",
-      "FOR SHARE OF package_versions",
-    ].join(" "),
-    [packageVersionId],
-  );
-  const row = result.rows[0];
-  if (row === undefined) {
-    throw new HttpError(
-      404,
-      `Catalog package version not found. packageVersionId=${packageVersionId}`,
-      "CATALOG_PACKAGE_VERSION_NOT_FOUND",
-    );
-  }
-
-  assertCatalogPackageVersionIsPublished(row);
-  return row;
-}
-
-async function countCatalogPackageVersionMediaAssetsInExecutor(
-  executor: DatabaseExecutor,
-  packageVersionId: string,
-): Promise<number> {
-  const result = await executor.query<Readonly<{ media_asset_count: string | number }>>(
-    [
-      "SELECT COUNT(*) AS media_asset_count",
-      "FROM catalog.package_media_assets",
-      "WHERE package_version_id = $1",
-    ].join(" "),
-    [packageVersionId],
-  );
-  const row = result.rows[0];
-  if (row === undefined) {
-    throw new Error("Catalog package version media asset count returned no rows");
-  }
-
-  return toSafeNumber(row.media_asset_count, "media_asset_count");
-}
-
-async function loadCatalogPackageVersionMediaAssetsInExecutor(
-  executor: DatabaseExecutor,
-  packageVersionId: string,
-): Promise<ReadonlyArray<CatalogPackageInstallMediaAssetRow>> {
-  const result = await executor.query<CatalogPackageInstallMediaAssetRow>(
-    [
-      "SELECT package_media_asset_id, package_media_key, media_blob_id",
-      "FROM catalog.package_media_assets",
-      "WHERE package_version_id = $1",
-      "ORDER BY package_media_key ASC",
-    ].join(" "),
-    [packageVersionId],
-  );
-
-  return result.rows;
-}
-
-async function loadCatalogPackageVersionCardsInExecutor(
-  executor: DatabaseExecutor,
-  packageVersionId: string,
-): Promise<ReadonlyArray<CatalogPackageInstallCardRow>> {
-  const result = await executor.query<CatalogPackageInstallCardRow>(
-    [
-      "SELECT package_card_id, stable_card_key, ordinal, front_text, back_text, card_type, metadata, tags, media_asset_keys",
-      "FROM catalog.package_cards",
-      "WHERE package_version_id = $1",
-      "ORDER BY ordinal ASC, package_card_id ASC",
-    ].join(" "),
-    [packageVersionId],
-  );
-
-  return result.rows;
-}
-
-async function loadCatalogPackageVersionTagRowsInExecutor(
-  executor: DatabaseExecutor,
-  packageVersionId: string,
-): Promise<ReadonlyArray<CatalogPackageInstallTagRow>> {
-  const result = await executor.query<CatalogPackageInstallTagRow>(
-    [
-      "SELECT tags",
-      "FROM catalog.package_cards",
-      "WHERE package_version_id = $1",
-      "ORDER BY ordinal ASC, package_card_id ASC",
-    ].join(" "),
-    [packageVersionId],
-  );
-
-  return result.rows;
-}
-
-function createCatalogPackageInstallTagPlan(
-  cards: ReadonlyArray<Readonly<{ tags: ReadonlyArray<string> }>>,
-  options: Readonly<{
-    addImportTag: boolean;
-    importTag: string;
-    removeTags: ReadonlyArray<string>;
-  }>,
-): CardImportTagPlan {
-  try {
-    return planCardImportTags(cards, options);
-  } catch (error) {
-    throw new HttpError(
-      400,
-      `Catalog package install tag options are invalid. reason=${error instanceof Error ? error.message : String(error)}`,
-      "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-    );
-  }
-}
-
-function createCatalogPackageInstallPreview(
-  row: CatalogPackageInstallVersionRow,
-  mediaAssetCount: number,
-  tagRows: ReadonlyArray<CatalogPackageInstallTagRow>,
-  input: CatalogPackageInstallPreviewInput,
-): CatalogPackageInstallPreview {
-  const suggestedImportTag = buildSuggestedCardImportTag(
-    input.generatedAt,
-    input.existingWorkspaceTags,
-  );
-  const tagPlan = createCatalogPackageInstallTagPlan(
-    tagRows,
-    {
-      addImportTag: true,
-      importTag: suggestedImportTag,
-      removeTags: [],
-    },
-  );
-
-  return {
-    packageVersion: mapCatalogPackageInstallPackageVersion(row),
-    summary: {
-      cardCount: toSafeNumber(row.card_count, "card_count"),
-      mediaAssetCount,
-    },
-    tagCounts: tagPlan.sourceTagCounts,
-    defaultOptions: {
-      addImportTag: true,
-      suggestedImportTag,
-      keptTags: tagPlan.keptTags,
-      removedTags: tagPlan.removedTags,
-    },
-  };
-}
-
-function buildCatalogInstallMediaOperationId(operationIdPrefix: string, mediaAssetIndex: number): string {
-  const lastOperationId = `${operationIdPrefix}:media:${mediaAssetIndex}`;
-  if (isValidMediaAssetLastOperationId(lastOperationId)) {
-    return lastOperationId;
-  }
-
-  throw new HttpError(
-    400,
-    "Derived catalog media lastOperationId is invalid.",
-    "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-  );
-}
-
-function buildCatalogInstallCardOperationId(operationIdPrefix: string, cardIndex: number): string {
-  const lastOperationId = `${operationIdPrefix}:card:${cardIndex}`;
-  if (isValidMediaAssetLastOperationId(lastOperationId)) {
-    return lastOperationId;
-  }
-
-  throw new HttpError(
-    400,
-    "Derived catalog card lastOperationId is invalid.",
-    "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-  );
-}
-
-function buildCatalogInstallOperationIds(
-  input: NormalizedCatalogPackageInstallConfirmInput,
-  mediaAssets: ReadonlyArray<CatalogPackageInstallMediaAssetRow>,
-  cards: ReadonlyArray<CatalogPackageInstallCardRow>,
-): ReadonlyArray<string> {
-  return [
-    ...mediaAssets.map((_mediaAsset, mediaAssetIndex) => (
-      buildCatalogInstallMediaOperationId(input.operationIdPrefix, mediaAssetIndex)
-    )),
-    ...cards.map((_card, cardIndex) => buildCatalogInstallCardOperationId(input.operationIdPrefix, cardIndex)),
-  ];
-}
-
-async function assertInstallIdUnusedInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  installId: string,
-): Promise<void> {
-  const result = await executor.query<Readonly<{ card_id: string }>>(
-    [
-      "SELECT card_id",
-      "FROM content.cards",
-      "WHERE workspace_id = $1",
-      "AND metadata->'source'->>'importId' = $2",
-      "LIMIT 1",
-    ].join(" "),
-    [workspaceId, installId],
-  );
-
-  const row = result.rows[0];
-  if (row === undefined) {
-    return;
-  }
-
-  throw new HttpError(
-    409,
-    `Catalog package install id already exists in this workspace. workspaceId=${workspaceId} installId=${installId} cardId=${row.card_id}`,
-    "CATALOG_PACKAGE_INSTALL_ID_ALREADY_EXISTS",
-  );
-}
-
-async function assertInstallOperationIdsUnusedInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  operationIds: ReadonlyArray<string>,
-): Promise<void> {
-  if (operationIds.length === 0) {
-    return;
-  }
-
-  const result = await executor.query<CatalogPackageInstallOperationConflictRow>(
-    [
-      "SELECT entity_type, entity_id, last_operation_id",
-      "FROM (",
-      "SELECT 'card'::text AS entity_type, card_id::text AS entity_id, last_operation_id",
-      "FROM content.cards",
-      "WHERE workspace_id = $1",
-      "AND last_operation_id = ANY($2::text[])",
-      "UNION ALL",
-      "SELECT 'media_asset'::text AS entity_type, media_asset_id::text AS entity_id, last_operation_id",
-      "FROM content.media_assets",
-      "WHERE workspace_id = $1",
-      "AND last_operation_id = ANY($2::text[])",
-      ") AS operation_conflicts",
-      "LIMIT 1",
-    ].join(" "),
-    [workspaceId, operationIds],
-  );
-
-  const row = result.rows[0];
-  if (row === undefined) {
-    return;
-  }
-
-  throw new HttpError(
-    409,
-    [
-      "Catalog package install operation id already exists in this workspace.",
-      `workspaceId=${workspaceId}`,
-      `operationId=${row.last_operation_id}`,
-      `entityType=${row.entity_type}`,
-      `entityId=${row.entity_id}`,
-    ].join(" "),
-    "CATALOG_PACKAGE_INSTALL_OPERATION_ALREADY_EXISTS",
-  );
-}
-
-async function assertCatalogInstallReplicaBelongsToWorkspaceInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  replicaId: string,
-): Promise<void> {
-  try {
-    await assertReplicaBelongsToWorkspaceInExecutor(executor, workspaceId, replicaId);
-  } catch (error) {
-    if (error instanceof HttpError && error.code === "MEDIA_ASSET_REPLICA_INVALID") {
-      throw new HttpError(
-        400,
-        "lastModifiedByReplicaId must reference a workspace replica for this workspace.",
-        "CATALOG_PACKAGE_INSTALL_REPLICA_INVALID",
-        error.details ?? undefined,
-      );
-    }
-
-    throw error;
-  }
-}
-
-function resolveInstalledMediaAssetId(
-  installedMediaAssetIdsByPackageMediaKey: ReadonlyMap<string, string>,
-  packageMediaKey: string,
-): string {
-  const mediaAssetId = installedMediaAssetIdsByPackageMediaKey.get(packageMediaKey);
-  if (mediaAssetId === undefined) {
-    throw new HttpError(
-      409,
-      `Catalog package card references missing package media asset. packageMediaKey=${packageMediaKey}`,
-      "CATALOG_PACKAGE_INSTALL_MEDIA_ASSET_NOT_FOUND",
-    );
-  }
-
-  return mediaAssetId;
-}
-
-function rewriteCatalogInstallMarkdown(
-  markdown: string,
-  installedMediaAssetIdsByPackageMediaKey: ReadonlyMap<string, string>,
-  packageCardId: string,
-  fieldName: "frontText" | "backText",
-): string {
-  try {
-    return rewriteMarkdownFcAssetUrlsToFcAssets(markdown, (packageMediaKey) => (
-      resolveInstalledMediaAssetId(installedMediaAssetIdsByPackageMediaKey, packageMediaKey)
-    ));
-  } catch (error) {
-    if (error instanceof HttpError) {
-      throw new HttpError(
-        error.statusCode,
-        `Catalog package card ${fieldName} media rewrite failed. packageCardId=${packageCardId} reason=${error.message}`,
-        error.code ?? undefined,
-        error.details ?? undefined,
-      );
-    }
-
-    throw new HttpError(
-      409,
-      `Catalog package card ${fieldName} media rewrite failed. packageCardId=${packageCardId} reason=${error instanceof Error ? error.message : String(error)}`,
-      "CATALOG_PACKAGE_INSTALL_MEDIA_REWRITE_FAILED",
-    );
-  }
-}
-
-function normalizePackageCardMetadata(card: CatalogPackageInstallCardRow): CardMetadata {
-  try {
-    return normalizeCardMetadata(card.metadata);
-  } catch (error) {
-    throw new HttpError(
-      409,
-      `Catalog package card metadata is invalid. packageCardId=${card.package_card_id} reason=${error instanceof Error ? error.message : String(error)}`,
-      "CATALOG_PACKAGE_CARD_METADATA_INVALID",
-    );
-  }
-}
-
-function normalizePackageCardSourceCreatedAt(
-  card: CatalogPackageInstallCardRow,
-  sourceCreatedAt: string | null,
-  packageSourceCreatedAt: string,
-): string {
-  if (sourceCreatedAt === null) {
-    return packageSourceCreatedAt;
-  }
-
-  try {
-    return normalizeIsoTimestamp(sourceCreatedAt, "metadata.source.createdAt");
-  } catch (error) {
-    throw new HttpError(
-      409,
-      `Catalog package card source createdAt is invalid. packageCardId=${card.package_card_id} reason=${error instanceof Error ? error.message : String(error)}`,
-      "CATALOG_PACKAGE_CARD_METADATA_INVALID",
-    );
-  }
-}
-
-function createCatalogPackageInstallPackageSource(
-  versionRow: CatalogPackageInstallVersionRow,
-): CardSourceMetadata & Readonly<{ createdAt: string }> {
-  return {
-    label: versionRow.title,
-    author: versionRow.author_display_name,
-    comment: versionRow.summary,
-    createdAt: toOptionalIsoString(versionRow.published_at) ?? toIsoString(versionRow.created_at),
-    importedAt: null,
-    importId: null,
-  };
-}
-
-function createCatalogPackageInstallCardMetadata(
-  card: CatalogPackageInstallCardRow,
-  versionRow: CatalogPackageInstallVersionRow,
-  input: NormalizedCatalogPackageInstallConfirmInput,
-): CardMetadata {
-  const cardMetadata = normalizePackageCardMetadata(card);
-  const packageSource = createCatalogPackageInstallPackageSource(versionRow);
-  const cardSource = cardMetadata.source;
-
-  return {
-    version: 1,
-    source: {
-      label: cardSource?.label ?? packageSource.label,
-      author: cardSource?.author ?? packageSource.author,
-      comment: cardSource?.comment ?? packageSource.comment,
-      createdAt: normalizePackageCardSourceCreatedAt(
-        card,
-        cardSource?.createdAt ?? null,
-        packageSource.createdAt,
-      ),
-      importedAt: input.installedAt,
-      importId: input.installId,
-    },
-  };
-}
-
-function assertPackageCardMediaAssetKeysExist(
-  card: CatalogPackageInstallCardRow,
-  installedMediaAssetIdsByPackageMediaKey: ReadonlyMap<string, string>,
-): void {
-  const missingMediaAssetKeys = card.media_asset_keys.filter((packageMediaKey) => (
-    installedMediaAssetIdsByPackageMediaKey.has(packageMediaKey) === false
-  ));
-  if (missingMediaAssetKeys.length === 0) {
-    return;
-  }
-
-  throw new HttpError(
-    409,
-    `Catalog package card references missing package media asset keys. packageCardId=${card.package_card_id} packageMediaKeys=${missingMediaAssetKeys.join(",")}`,
-    "CATALOG_PACKAGE_INSTALL_MEDIA_ASSET_NOT_FOUND",
-  );
-}
-
-async function insertWorkspaceMediaAssetForCatalogInstallInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  mediaAssetId: string,
-  packageMediaAsset: CatalogPackageInstallMediaAssetRow,
-  input: NormalizedCatalogPackageInstallConfirmInput,
-  mediaAssetIndex: number,
-): Promise<void> {
-  await executor.query(
-    [
-      "INSERT INTO content.media_assets",
-      "(",
-      "media_asset_id, workspace_id, media_blob_id, source_url, created_at,",
-      "client_updated_at, last_modified_by_replica_id, last_operation_id",
-      ")",
-      "VALUES ($1, $2, $3, NULL, $4, $5, $6, $7)",
-    ].join(" "),
-    [
-      mediaAssetId,
-      workspaceId,
-      packageMediaAsset.media_blob_id,
-      input.installedAt,
-      input.clientUpdatedAt,
-      input.lastModifiedByReplicaId,
-      buildCatalogInstallMediaOperationId(input.operationIdPrefix, mediaAssetIndex),
-    ],
-  );
-}
-
-async function insertWorkspaceCardForCatalogInstallInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  cardId: string,
-  card: CatalogPackageInstallCardRow,
-  input: NormalizedCatalogPackageInstallConfirmInput,
-  metadata: CardMetadata,
-  tags: ReadonlyArray<string>,
-  createdAt: string,
-  cardIndex: number,
-  installedMediaAssetIdsByPackageMediaKey: ReadonlyMap<string, string>,
-): Promise<void> {
-  assertPackageCardMediaAssetKeysExist(card, installedMediaAssetIdsByPackageMediaKey);
-
-  await executor.query(
-    [
-      "INSERT INTO content.cards",
-      "(",
-      "card_id, workspace_id, front_text, back_text, card_type, metadata, tags, effort_level, due_at, created_at,",
-      "reps, lapses, fsrs_card_state, fsrs_step_index, fsrs_stability, fsrs_difficulty, fsrs_last_reviewed_at, fsrs_scheduled_days,",
-      "client_updated_at, last_modified_by_replica_id, last_operation_id",
-      ")",
-      "VALUES ($1, $2, $3, $4, $5, $6::jsonb, $7, 'fast', NULL, $8, 0, 0, 'new', NULL, NULL, NULL, NULL, NULL, $9, $10, $11)",
-    ].join(" "),
-    [
-      cardId,
-      workspaceId,
-      rewriteCatalogInstallMarkdown(
-        card.front_text,
-        installedMediaAssetIdsByPackageMediaKey,
-        card.package_card_id,
-        "frontText",
-      ),
-      rewriteCatalogInstallMarkdown(
-        card.back_text,
-        installedMediaAssetIdsByPackageMediaKey,
-        card.package_card_id,
-        "backText",
-      ),
-      card.card_type,
-      JSON.stringify(metadata),
-      tags,
-      createdAt,
-      input.clientUpdatedAt,
-      input.lastModifiedByReplicaId,
-      buildCatalogInstallCardOperationId(input.operationIdPrefix, cardIndex),
-    ],
-  );
-}
-
-async function installCatalogPackageMediaAssetsInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  hotChangeWriteLock: HotChangeWriteLock,
-  mediaAssets: ReadonlyArray<CatalogPackageInstallMediaAssetRow>,
-  input: NormalizedCatalogPackageInstallConfirmInput,
-): Promise<ReadonlyArray<CatalogInstalledMediaAsset>> {
-  const installedMediaAssets: Array<CatalogInstalledMediaAsset> = [];
-
-  for (const [mediaAssetIndex, packageMediaAsset] of mediaAssets.entries()) {
-    const mediaAssetId = randomUUID();
-    await insertWorkspaceMediaAssetForCatalogInstallInExecutor(
-      executor,
-      workspaceId,
-      mediaAssetId,
-      packageMediaAsset,
-      input,
-      mediaAssetIndex,
-    );
-    await insertSyncChange(
-      executor,
-      workspaceId,
-      hotChangeWriteLock,
-      "media_asset",
-      mediaAssetId,
-      "upsert",
-      input.lastModifiedByReplicaId,
-      buildCatalogInstallMediaOperationId(input.operationIdPrefix, mediaAssetIndex),
-      input.clientUpdatedAt,
-    );
-
-    installedMediaAssets.push({
-      packageMediaAssetId: packageMediaAsset.package_media_asset_id,
-      packageMediaKey: packageMediaAsset.package_media_key,
-      mediaAssetId,
-    });
-  }
-
-  return installedMediaAssets;
-}
-
-function buildInstalledMediaAssetIdsByPackageMediaKey(
-  installedMediaAssets: ReadonlyArray<CatalogInstalledMediaAsset>,
-): ReadonlyMap<string, string> {
-  return new Map(installedMediaAssets.map((mediaAsset) => [
-    mediaAsset.packageMediaKey,
-    mediaAsset.mediaAssetId,
-  ]));
-}
-
-function buildOrderedCatalogInstallCardTimestamps(
-  installedAt: string,
-  cardCount: number,
-): ReadonlyArray<string> {
-  const installedAtMillis = new Date(installedAt).getTime();
-  return Array.from({ length: cardCount }, (_unused, cardIndex) => {
-    const timestampMillis = installedAtMillis - (cardCount - cardIndex - 1);
-    try {
-      return new Date(timestampMillis).toISOString();
-    } catch {
-      throw new HttpError(
-        400,
-        `installedAt is too early to assign ordered catalog card timestamps. installedAt=${installedAt} cardCount=${cardCount}`,
-        "CATALOG_PACKAGE_INSTALL_INVALID_INPUT",
-      );
-    }
-  });
-}
-
-function getCatalogInstallCardValue<Value>(
-  values: ReadonlyArray<Value>,
-  cardIndex: number,
-  fieldName: string,
-): Value {
-  const value = values[cardIndex];
-  if (value === undefined) {
-    throw new Error(`Missing catalog install card value. fieldName=${fieldName} cardIndex=${cardIndex}`);
-  }
-
-  return value;
-}
-
-async function installCatalogPackageCardsInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  hotChangeWriteLock: HotChangeWriteLock,
-  versionRow: CatalogPackageInstallVersionRow,
-  cards: ReadonlyArray<CatalogPackageInstallCardRow>,
-  tagPlan: CardImportTagPlan,
-  input: NormalizedCatalogPackageInstallConfirmInput,
-  installedMediaAssetIdsByPackageMediaKey: ReadonlyMap<string, string>,
-): Promise<ReadonlyArray<CatalogInstalledCard>> {
-  const installedCards: Array<CatalogInstalledCard> = [];
-  const createdAtValues = buildOrderedCatalogInstallCardTimestamps(input.installedAt, cards.length);
-
-  for (const [cardIndex, card] of cards.entries()) {
-    const cardId = randomUUID();
-    await insertWorkspaceCardForCatalogInstallInExecutor(
-      executor,
-      workspaceId,
-      cardId,
-      card,
-      input,
-      createCatalogPackageInstallCardMetadata(card, versionRow, input),
-      getCatalogInstallCardValue(tagPlan.cardTags, cardIndex, "tags"),
-      getCatalogInstallCardValue(createdAtValues, cardIndex, "createdAt"),
-      cardIndex,
-      installedMediaAssetIdsByPackageMediaKey,
-    );
-    await insertSyncChange(
-      executor,
-      workspaceId,
-      hotChangeWriteLock,
-      "card",
-      cardId,
-      "upsert",
-      input.lastModifiedByReplicaId,
-      buildCatalogInstallCardOperationId(input.operationIdPrefix, cardIndex),
-      input.clientUpdatedAt,
-    );
-    installedCards.push({
-      packageCardId: card.package_card_id,
-      stableCardKey: card.stable_card_key,
-      ordinal: toSafeNumber(card.ordinal, "ordinal"),
-      cardId,
-    });
-  }
-
-  return installedCards;
-}
-
-export async function previewCatalogPackageInstallInExecutor(
-  executor: DatabaseExecutor,
-  packageVersionId: string,
-  input: CatalogPackageInstallPreviewInput,
-): Promise<CatalogPackageInstallPreview> {
-  const normalizedPackageVersionId = normalizeUuidString(packageVersionId, "packageVersionId");
-  const versionRow = await loadCatalogPackageInstallVersionInExecutor(executor, normalizedPackageVersionId);
-  const mediaAssetCount = await countCatalogPackageVersionMediaAssetsInExecutor(
-    executor,
-    normalizedPackageVersionId,
-  );
-  const tagRows = await loadCatalogPackageVersionTagRowsInExecutor(
-    executor,
-    normalizedPackageVersionId,
-  );
-
-  return createCatalogPackageInstallPreview(versionRow, mediaAssetCount, tagRows, input);
-}
-
-export async function installCatalogPackageVersionInExecutor(
-  executor: DatabaseExecutor,
-  workspaceId: string,
-  packageVersionId: string,
-  input: CatalogPackageInstallConfirmInput,
-): Promise<CatalogPackageInstallResult> {
-  const normalizedWorkspaceId = normalizeCatalogWorkspaceId(workspaceId);
-  const normalizedPackageVersionId = normalizeUuidString(packageVersionId, "packageVersionId");
-  const normalizedInput = normalizeCatalogPackageInstallConfirmInput(input);
-  const requestIdentity = createCatalogPackageInstallRequestIdentity(
-    normalizedPackageVersionId,
-    normalizedInput,
-  );
-  const hotChangeWriteLock = await lockWorkspaceSyncMetadataForHotChangesInExecutor(
-    executor,
-    normalizedWorkspaceId,
-  );
-  const replayResult = await loadCatalogPackageInstallReplayInExecutor(
-    executor,
-    normalizedWorkspaceId,
-    normalizedInput.installId,
-    requestIdentity,
-  );
-  if (replayResult !== null) {
-    return replayResult;
-  }
-
-  const versionRow = await loadCatalogPackageInstallVersionForInstallInExecutor(
-    executor,
-    normalizedPackageVersionId,
-  );
-  const packageMediaAssets = await loadCatalogPackageVersionMediaAssetsInExecutor(
-    executor,
-    normalizedPackageVersionId,
-  );
-  const packageCards = await loadCatalogPackageVersionCardsInExecutor(executor, normalizedPackageVersionId);
-
-  if (packageCards.length === 0) {
-    throw new HttpError(
-      409,
-      `Catalog package version has no cards to install. packageVersionId=${normalizedPackageVersionId}`,
-      "CATALOG_PACKAGE_VERSION_EMPTY",
-    );
-  }
-  const tagPlan = createCatalogPackageInstallTagPlan(packageCards, normalizedInput);
-
-  await assertCatalogInstallReplicaBelongsToWorkspaceInExecutor(
-    executor,
-    normalizedWorkspaceId,
-    normalizedInput.lastModifiedByReplicaId,
-  );
-  await assertInstallIdUnusedInExecutor(executor, normalizedWorkspaceId, normalizedInput.installId);
-  await assertInstallOperationIdsUnusedInExecutor(
-    executor,
-    normalizedWorkspaceId,
-    buildCatalogInstallOperationIds(normalizedInput, packageMediaAssets, packageCards),
-  );
-
-  const installedMediaAssets = await installCatalogPackageMediaAssetsInExecutor(
-    executor,
-    normalizedWorkspaceId,
-    hotChangeWriteLock,
-    packageMediaAssets,
-    normalizedInput,
-  );
-  const installedMediaAssetIdsByPackageMediaKey = buildInstalledMediaAssetIdsByPackageMediaKey(
-    installedMediaAssets,
-  );
-  const installedCards = await installCatalogPackageCardsInExecutor(
-    executor,
-    normalizedWorkspaceId,
-    hotChangeWriteLock,
-    versionRow,
-    packageCards,
-    tagPlan,
-    normalizedInput,
-    installedMediaAssetIdsByPackageMediaKey,
-  );
-
-  const result: CatalogPackageInstallResult = {
-    packageVersion: mapCatalogPackageInstallPackageVersion(versionRow),
-    installedCards,
-    installedMediaAssets,
-    summary: {
-      cardCount: installedCards.length,
-      mediaAssetCount: installedMediaAssets.length,
-      installId: normalizedInput.installId,
-      installedAt: normalizedInput.installedAt,
-      keptTagCount: tagPlan.keptTags.length,
-      removedTagCount: tagPlan.removedTags.length,
-      importTag: tagPlan.importTag,
-    },
-  };
-  await insertCatalogPackageInstallIdempotencyResultInExecutor(
-    executor,
-    normalizedWorkspaceId,
-    normalizedInput.installId,
-    requestIdentity,
-    result,
-  );
-  return result;
-}
-
-export async function previewCatalogPackageInstall(
-  userId: string,
-  workspaceId: string,
-  packageVersionId: string,
-  input: CatalogPackageInstallPreviewInput,
-): Promise<CatalogPackageInstallPreview> {
-  return transactionWithWorkspaceScopeReadOnly({ userId, workspaceId }, async (executor) => (
-    previewCatalogPackageInstallInExecutor(executor, packageVersionId, input)
-  ));
-}
-
-export async function installCatalogPackageVersion(
-  userId: string,
-  workspaceId: string,
-  packageVersionId: string,
-  input: CatalogPackageInstallConfirmInput,
-): Promise<CatalogPackageInstallResult> {
-  return transactionWithWorkspaceScope({ userId, workspaceId }, async (executor) => (
-    installCatalogPackageVersionInExecutor(executor, workspaceId, packageVersionId, input)
-  ));
-}
