+++ apps/backend/src/catalog/distribution/install/persistence.test.ts
@@ -2,18 +2,15 @@ import assert from "node:assert/strict";
 import test from "node:test";
 import type pg from "pg";
 import type { DatabaseExecutor, SqlValue } from "../../../database";
-import { createCatalogInstallRoutes } from "../../../routes/catalogInstall";
 import { HttpError } from "../../../shared/errors";
 import {
   catalogPackageInstallOperationIdPrefixMaximumLength,
   installCatalogPackageVersionInExecutor,
-  previewCatalogPackageInstallInExecutor,
-} from "./install";
+} from "./index";
 import {
-  createQueryResult,
-  testAuthorId,
+  createPackageInstallVersionRow,
+  testInstallTimestamp,
   testMediaBlobId,
-  testPackageId,
   testPackageMediaAssetId,
   testPackageMediaKey,
   testPackageVersionId,
@@ -21,289 +18,9 @@ import {
   testWorkspaceCardId,
   testWorkspaceId,
   testWorkspaceMediaAssetId,
-} from "../../testSupport";
-import type {
-  CatalogPackageInstallConfirmInput,
-  CatalogPackageInstallResult,
-  CatalogPackageStatus,
-} from "../../types";
-
-const testWorkspaceReplicaId = "bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb";
-const testInstallTimestamp = "2026-04-19T10:30:00.000Z";
-
-function createPackageInstallVersionRow(status: CatalogPackageStatus): Readonly<Record<string, unknown>> {
-  return {
-    package_version_id: testPackageVersionId,
-    package_id: testPackageId,
-    version_number: 1,
-    status,
-    slug: "spanish-basics",
-    title: "Spanish Basics",
-    summary: "Core Spanish prompts.",
-    description: "Core Spanish flashcards for beginners.",
-    language_tags: ["en", "es"],
-    topic_tags: ["language"],
-    license: "CC-BY-4.0",
-    content_warning: null,
-    cover_package_media_key: null,
-    card_count: 1,
-    created_at: testTimestamp,
-    published_at: testTimestamp,
-    author_id: testAuthorId,
-    author_slug: "open-cards",
-    author_display_name: "Open Cards",
-  };
-}
-
-function createStoredCatalogPackageInstallResult(installId: string): CatalogPackageInstallResult {
-  return {
-    packageVersion: {
-      packageVersionId: testPackageVersionId,
-      packageId: testPackageId,
-      versionNumber: 1,
-      slug: "spanish-basics",
-      title: "Spanish Basics",
-      summary: "Core Spanish prompts.",
-      description: "Core Spanish flashcards for beginners.",
-      languageTags: ["en", "es"],
-      topicTags: ["language"],
-      license: "CC-BY-4.0",
-      contentWarning: null,
-      coverPackageMediaKey: null,
-      cardCount: 1,
-      createdAt: testTimestamp,
-      publishedAt: testTimestamp,
-      author: {
-        authorId: testAuthorId,
-        slug: "open-cards",
-        displayName: "Open Cards",
-      },
-    },
-    installedCards: [{
-      packageCardId: testWorkspaceCardId,
-      stableCardKey: "hola-card",
-      ordinal: 1,
-      cardId: testWorkspaceCardId,
-    }],
-    installedMediaAssets: [{
-      packageMediaAssetId: testPackageMediaAssetId,
-      packageMediaKey: testPackageMediaKey,
-      mediaAssetId: testWorkspaceMediaAssetId,
-    }],
-    summary: {
-      cardCount: 1,
-      mediaAssetCount: 1,
-      installId,
-      installedAt: testInstallTimestamp,
-      keptTagCount: 1,
-      removedTagCount: 0,
-      importTag: null,
-    },
-  };
-}
-
-function createCatalogPackageInstallReplayExecutor(
-  input: CatalogPackageInstallConfirmInput,
-  installResult: unknown,
-): DatabaseExecutor {
-  return {
-    async query<Row extends pg.QueryResultRow>(
-      text: string,
-      params: ReadonlyArray<SqlValue>,
-    ): Promise<pg.QueryResult<Row>> {
-      if (text.includes("INSERT INTO sync.workspace_sync_metadata")) {
-        assert.deepEqual(params, [testWorkspaceId]);
-        return createQueryResult([]);
-      }
-      if (text.includes("FROM sync.workspace_sync_metadata") && text.includes("FOR UPDATE")) {
-        assert.deepEqual(params, [testWorkspaceId]);
-        return createQueryResult([{ workspace_id: testWorkspaceId } as unknown as Row]);
-      }
-      if (text.includes("FROM sync.catalog_package_install_idempotency")) {
-        assert.deepEqual(params, [testWorkspaceId, input.installId]);
-        return createQueryResult([{
-          package_version_id: testPackageVersionId,
-          installed_at: testInstallTimestamp,
-          client_updated_at: testInstallTimestamp,
-          last_modified_by_replica_id: testWorkspaceReplicaId,
-          operation_id_prefix: input.operationIdPrefix,
-          add_import_tag: false,
-          import_tag: null,
-          remove_tags: [],
-          install_result: installResult,
-        } as unknown as Row]);
-      }
-
-      throw new Error(`Unexpected query: ${text}`);
-    },
-  };
-}
-
-test("catalog install preview exposes source tag counts and ZIP-compatible default options", async () => {
-  const executor: DatabaseExecutor = {
-    async query<Row extends pg.QueryResultRow>(
-      text: string,
-      params: ReadonlyArray<SqlValue>,
-    ): Promise<pg.QueryResult<Row>> {
-      assert.deepEqual(params, [testPackageVersionId]);
-      if (text.includes("FROM catalog.package_versions AS package_versions")) {
-        return createQueryResult([{
-          ...createPackageInstallVersionRow("published"),
-          card_count: 2,
-        } as unknown as Row]);
-      }
-      if (text.includes("COUNT(*) AS media_asset_count")) {
-        return createQueryResult([{ media_asset_count: 0 } as unknown as Row]);
-      }
-      if (text.includes("SELECT tags") && text.includes("FROM catalog.package_cards")) {
-        assert.match(text, /ORDER BY ordinal ASC, package_card_id ASC/);
-        return createQueryResult([
-          { tags: ["language", "legacy"] } as unknown as Row,
-          { tags: ["language"] } as unknown as Row,
-        ]);
-      }
-
-      throw new Error(`Unexpected query: ${text}`);
-    },
-  };
-
-  const preview = await previewCatalogPackageInstallInExecutor(
-    executor,
-    testPackageVersionId,
-    {
-      generatedAt: testInstallTimestamp,
-      existingWorkspaceTags: ["", "import:2026-04-19-0"],
-    },
-  );
-
-  assert.deepEqual(preview.tagCounts, [
-    { tag: "language", cardsCount: 2 },
-    { tag: "legacy", cardsCount: 1 },
-  ]);
-  assert.deepEqual(preview.defaultOptions, {
-    addImportTag: true,
-    suggestedImportTag: "import:2026-04-19-1",
-    keptTags: ["language", "legacy"],
-    removedTags: [],
-  });
-});
-
-test("catalog install rejects stored replay results that violate the published response contract", async () => {
-  const installInput: CatalogPackageInstallConfirmInput = {
-    installId: "catalog-install-stored-contract",
-    installedAt: testInstallTimestamp,
-    clientUpdatedAt: testInstallTimestamp,
-    lastModifiedByReplicaId: testWorkspaceReplicaId,
-    operationIdPrefix: "catalog-install-stored-contract",
-  };
-  const validResult = createStoredCatalogPackageInstallResult(installInput.installId);
-  const invalidResults: ReadonlyArray<Readonly<{ name: string; result: unknown }>> = [
-    {
-      name: "invalid UUID",
-      result: {
-        ...validResult,
-        packageVersion: {
-          ...validResult.packageVersion,
-          author: { ...validResult.packageVersion.author, authorId: "not-a-uuid" },
-        },
-      },
-    },
-    {
-      name: "invalid date-time",
-      result: {
-        ...validResult,
-        packageVersion: { ...validResult.packageVersion, createdAt: "not-a-date-time" },
-      },
-    },
-    {
-      name: "nonpositive version number",
-      result: {
-        ...validResult,
-        packageVersion: { ...validResult.packageVersion, versionNumber: 0 },
-      },
-    },
-    {
-      name: "negative response count",
-      result: {
-        ...validResult,
-        summary: { ...validResult.summary, cardCount: -1 },
-      },
-    },
-  ];
-
-  for (const invalidResult of invalidResults) {
-    await assert.rejects(
-      installCatalogPackageVersionInExecutor(
-        createCatalogPackageInstallReplayExecutor(installInput, invalidResult.result),
-        testWorkspaceId,
-        testPackageVersionId,
-        installInput,
-      ),
-      (error: unknown): boolean => {
-        assert.ok(error instanceof HttpError);
-        assert.equal(error.statusCode, 500);
-        assert.equal(error.code, "CATALOG_PACKAGE_INSTALL_STORED_RESULT_INVALID");
-        assert.match(error.message, /cannot be replayed safely/);
-        assert.match(error.message, /Repair the catalog install idempotency record/);
-        return true;
-      },
-      invalidResult.name,
-    );
-  }
-});
-
-test("catalog install rejects a contract-valid stored result that mismatches its durable identity", async () => {
-  const installInput: CatalogPackageInstallConfirmInput = {
-    installId: "catalog-install-stored-identity",
-    installedAt: testInstallTimestamp,
-    clientUpdatedAt: testInstallTimestamp,
-    lastModifiedByReplicaId: testWorkspaceReplicaId,
-    operationIdPrefix: "catalog-install-stored-identity",
-  };
-  const validResult = createStoredCatalogPackageInstallResult(installInput.installId);
-  const mismatchedResult: CatalogPackageInstallResult = {
-    ...validResult,
-    packageVersion: {
-      ...validResult.packageVersion,
-      packageVersionId: "99999999-9999-4999-8999-999999999999",
-      cardCount: 2,
-    },
-    summary: {
-      ...validResult.summary,
-      installId: "different-install-id",
-      installedAt: "2026-04-19T10:30:01.000Z",
-      cardCount: 2,
-      mediaAssetCount: 0,
-      removedTagCount: 1,
-      importTag: "different-import-tag",
-    },
-  };
-
-  await assert.rejects(
-    installCatalogPackageVersionInExecutor(
-      createCatalogPackageInstallReplayExecutor(installInput, mismatchedResult),
-      testWorkspaceId,
-      testPackageVersionId,
-      installInput,
-    ),
-    (error: unknown): boolean => {
-      assert.ok(error instanceof HttpError);
-      assert.equal(error.statusCode, 500);
-      assert.equal(error.code, "CATALOG_PACKAGE_INSTALL_STORED_RESULT_INVALID");
-      assert.match(error.message, /does not match its durable request identity/);
-      assert.match(error.message, /summary\.installId/);
-      assert.match(error.message, /packageVersion\.packageVersionId/);
-      assert.match(error.message, /summary\.installedAt/);
-      assert.match(error.message, /summary\.importTag/);
-      assert.match(error.message, /summary\.cardCount/);
-      assert.match(error.message, /packageVersion\.cardCount/);
-      assert.match(error.message, /summary\.mediaAssetCount/);
-      assert.match(error.message, /summary\.removedTagCount/);
-      assert.match(error.message, /Repair the catalog install idempotency record/);
-      return true;
-    },
-  );
-});
+  testWorkspaceReplicaId,
+} from "./installTestSupport";
+import { createQueryResult } from "../../testSupport";
 
 test("catalog install creates logical media assets from existing blobs and rewrites card markdown", async () => {
   const queries: Array<Readonly<{ text: string; params: ReadonlyArray<SqlValue> }>> = [];
@@ -939,201 +656,4 @@ test("catalog install rejects invalid package card source createdAt", async () =
   assert.equal(queries.some((query) => query.text.includes("INSERT INTO sync.hot_changes")), false);
 });
 
-test("catalog install rejects unpublished and delisted package versions", async () => {
-  const previewExecutor: DatabaseExecutor = {
-    async query<Row extends pg.QueryResultRow>(
-      text: string,
-      params: ReadonlyArray<SqlValue>,
-    ): Promise<pg.QueryResult<Row>> {
-      assert.match(text, /FROM catalog\.package_versions AS package_versions/);
-      assert.deepEqual(params, [testPackageVersionId]);
-      return createQueryResult([createPackageInstallVersionRow("draft") as unknown as Row]);
-    },
-  };
 
-  await assert.rejects(
-    previewCatalogPackageInstallInExecutor(
-      previewExecutor,
-      testPackageVersionId,
-      {
-        generatedAt: testInstallTimestamp,
-        existingWorkspaceTags: [],
-      },
-    ),
-    (error: unknown) => {
-      assert.equal(error instanceof HttpError, true);
-      assert.equal((error as HttpError).statusCode, 409);
-      assert.equal((error as HttpError).code, "CATALOG_PACKAGE_VERSION_NOT_PUBLISHED");
-      assert.match((error as HttpError).message, /status=draft/);
-      return true;
-    },
-  );
-
-  const installExecutor: DatabaseExecutor = {
-    async query<Row extends pg.QueryResultRow>(
-      text: string,
-      params: ReadonlyArray<SqlValue>,
-    ): Promise<pg.QueryResult<Row>> {
-      if (text.includes("INSERT INTO sync.workspace_sync_metadata")) {
-        return createQueryResult([]);
-      }
-      if (text.includes("FROM sync.workspace_sync_metadata") && text.includes("FOR UPDATE")) {
-        return createQueryResult([{ workspace_id: testWorkspaceId } as unknown as Row]);
-      }
-      if (text.includes("FROM sync.catalog_package_install_idempotency")) {
-        return createQueryResult([]);
-      }
-      assert.match(text, /FROM catalog\.package_versions AS package_versions/);
-      assert.match(text, /FOR SHARE OF package_versions/);
-      assert.deepEqual(params, [testPackageVersionId]);
-      return createQueryResult([createPackageInstallVersionRow("delisted") as unknown as Row]);
-    },
-  };
-
-  await assert.rejects(
-    installCatalogPackageVersionInExecutor(
-      installExecutor,
-      testWorkspaceId,
-      testPackageVersionId,
-      {
-        installId: "catalog-install-2",
-        installedAt: testInstallTimestamp,
-        clientUpdatedAt: testInstallTimestamp,
-        lastModifiedByReplicaId: testWorkspaceReplicaId,
-        operationIdPrefix: "catalog-install-2",
-      },
-    ),
-    (error: unknown) => {
-      assert.equal(error instanceof HttpError, true);
-      assert.equal((error as HttpError).statusCode, 409);
-      assert.equal((error as HttpError).code, "CATALOG_PACKAGE_VERSION_NOT_PUBLISHED");
-      assert.match((error as HttpError).message, /status=delisted/);
-      return true;
-    },
-  );
-});
-
-test("catalog install route rejects unauthorized workspace access before installing", async () => {
-  let previewCalled = false;
-  const app = createCatalogInstallRoutes({
-    allowedOrigins: [],
-    loadRequestContextFromRequestFn: async () => ({
-      requestAuthInputs: {
-        authorizationHeader: undefined,
-        sessionToken: undefined,
-        csrfTokenHeader: undefined,
-        originHeader: undefined,
-        refererHeader: undefined,
-        secFetchSiteHeader: undefined,
-      },
-      requestContext: {
-        userId: "user-1",
-        subjectUserId: "subject-user-1",
-        selectedWorkspaceId: null,
-        email: "user@example.com",
-        locale: "en",
-        userSettingsCreatedAt: testTimestamp,
-        preferences: {
-          reviewReactionAnimationsEnabled: true,
-        },
-        transport: "api_key",
-        connectionId: "connection-1",
-        guestSessionId: null,
-        guestPlatform: null,
-      },
-    }),
-    assertUserHasWorkspaceAccessFn: async () => {
-      throw new HttpError(404, "Workspace not found", "WORKSPACE_NOT_FOUND");
-    },
-    previewCatalogPackageInstallFn: async () => {
-      previewCalled = true;
-      throw new Error("preview should not run");
-    },
-  });
-  app.onError((error) => {
-    throw error;
-  });
-
-  await assert.rejects(
-    async () => (
-      app.request(
-        `http://localhost/workspaces/${testWorkspaceId}/catalog/package-versions/${testPackageVersionId}/install/preview`,
-        { method: "POST" },
-      )
-    ),
-    (error: unknown) => {
-      assert.equal(error instanceof HttpError, true);
-      assert.equal((error as HttpError).statusCode, 404);
-      assert.equal((error as HttpError).code, "WORKSPACE_NOT_FOUND");
-      return true;
-    },
-  );
-  assert.equal(previewCalled, false);
-});
-
-test("catalog install route rejects unsafe operation prefixes without sanitizing them", async () => {
-  let installCalled = false;
-  const app = createCatalogInstallRoutes({
-    allowedOrigins: [],
-    loadRequestContextFromRequestFn: async () => ({
-      requestAuthInputs: {
-        authorizationHeader: undefined,
-        sessionToken: undefined,
-        csrfTokenHeader: undefined,
-        originHeader: undefined,
-        refererHeader: undefined,
-        secFetchSiteHeader: undefined,
-      },
-      requestContext: {
-        userId: "user-1",
-        subjectUserId: "subject-user-1",
-        selectedWorkspaceId: null,
-        email: "user@example.com",
-        locale: "en",
-        userSettingsCreatedAt: testTimestamp,
-        preferences: {
-          reviewReactionAnimationsEnabled: true,
-        },
-        transport: "api_key",
-        connectionId: "connection-1",
-        guestSessionId: null,
-        guestPlatform: null,
-      },
-    }),
-    assertUserHasWorkspaceAccessFn: async () => undefined,
-    installCatalogPackageVersionFn: async () => {
-      installCalled = true;
-      throw new Error("catalog install must not run for an unsafe operationIdPrefix");
-    },
-  });
-  app.onError((error) => {
-    throw error;
-  });
-
-  await assert.rejects(
-    async () => app.request(
-      `http://localhost/workspaces/${testWorkspaceId}/catalog/package-versions/${testPackageVersionId}/install`,
-      {
-        method: "POST",
-        headers: {
-          "content-type": "application/json",
-        },
-        body: JSON.stringify({
-          installId: "catalog-install-route-invalid-prefix",
-          installedAt: testInstallTimestamp,
-          clientUpdatedAt: testInstallTimestamp,
-          lastModifiedByReplicaId: testWorkspaceReplicaId,
-          operationIdPrefix: " trailing-space ",
-        }),
-      },
-    ),
-    (error: unknown): boolean => {
-      assert.ok(error instanceof HttpError);
-      assert.equal(error.statusCode, 400);
-      assert.equal(error.code, "CATALOG_PACKAGE_INSTALL_INVALID_INPUT");
-      assert.match(error.message, /operationIdPrefix.*printable ASCII/);
-      return true;
-    },
-  );
-  assert.equal(installCalled, false);
-});
